@@ -3,10 +3,10 @@ apply plugin: 'com.android.library'
 group='com.github.chuanchic'
 
 android {
-    compileSdkVersion 28
+    compileSdkVersion 29
     defaultConfig {
-        minSdkVersion 16
-        targetSdkVersion 28
+        minSdkVersion 19
+        targetSdkVersion 29
     }
 
     buildTypes {
@@ -18,6 +18,6 @@ android {
 }
 
 dependencies {
-    api 'com.android.support:recyclerview-v7:28.0.0'
-    api 'com.android.support:appcompat-v7:28.0.0'
+    api 'androidx.recyclerview:recyclerview:1.1.0'
+    api 'androidx.appcompat:appcompat:1.2.0'
 }
