@@ -1,7 +1,7 @@
 package com.github.chuanchic.helper;
 
 import android.app.Activity;
-import android.support.v4.app.Fragment;
+import androidx.fragment.app.Fragment;
 import android.util.DisplayMetrics;
 import android.view.View;
 import android.widget.AbsListView;
